# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gradovi API",
    "version": "1.0.0",
    "description": "API za pristup informacijama o gradovima",
    "contact": {
      "name": "Luka Miličević",
      "url": "https://github.com/Luka147m"
    },
    "license": {
      "name": "CC0 1.0 Universal",
      "url": "https://github.com/Luka147m/OR-labos/blob/main/LICENSE"
    }
  },
  "paths": {
    "/api/gradovi": {
      "get": {
        "summary": "Dohvaća sve gradove iz baze podataka",
        "responses": {
          "200": {
            "description": "Uspijeh u dohvaćanju podataka, vraćeni podatci su u JSON-u",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "status": "OK",
                  "message": "Fetched all entries from database",
                  "response": [
                    {
                      "id": 1,
                      "imegrada": "Zagreb",
                      "kvartovi": [
                        {
                          "nazivkvarta": "Gornja Dubrava",
                          "brojkvartstan": 58908
                        },
                        {
                          "nazivkvarta": "Maksimir",
                          "brojkvartstan": 47989
                        },
                        {
                          "nazivkvarta": "Trešnjevka",
                          "brojkvartstan": 119475
                        }
                      ],
                      "latitude": 45.813,
                      "povrsina": 641.2,
                      "zupanija": "Grad Zagreb",
                      "longitude": 15.9775,
                      "godinaosnutka": 1094,
                      "gradonacelnik": "Tomislav Tomašević",
                      "brojstanovnika": 777183,
                      "nadmorskavisina": 122,
                      "links": [
                        {
                          "href": "/api/grad/1",
                          "rel": "self",
                          "type": "GET"
                        },
                        {
                          "href": "/api/grad/ime/Zagreb",
                          "rel": "self",
                          "type": "GET"
                        },
                        {
                          "href": "/api/gradovi/populacija/777182",
                          "rel": "related",
                          "type": "GET"
                        }
                      ]
                    },
                    {
                      "id": 5,
                      "imegrada": "Slavonski Brod",
                      "kvartovi": [],
                      "latitude": 45.16,
                      "povrsina": 50.27,
                      "zupanija": "Brodsko-posavska",
                      "longitude": 18.01,
                      "godinaosnutka": 1224,
                      "gradonacelnik": "Mirko Duspara",
                      "brojstanovnika": 51264,
                      "nadmorskavisina": 96,
                      "links": [
                        {
                          "href": "/api/grad/5",
                          "rel": "self",
                          "type": "GET"
                        },
                        {
                          "href": "/api/grad/ime/Slavonski Brod",
                          "rel": "self",
                          "type": "GET"
                        },
                        {
                          "href": "/api/gradovi/populacija/51263",
                          "rel": "related",
                          "type": "GET"
                        }
                      ]
                    },
                    {
                      "id": 4,
                      "imegrada": "Osijek",
                      "kvartovi": [
                        {
                          "nazivkvarta": "Tvrđa",
                          "brojkvartstan": 10277
                        }
                      ],
                      "latitude": 45.551,
                      "povrsina": 171,
                      "zupanija": "Osječko-baranjska",
                      "longitude": 18.694,
                      "godinaosnutka": 124,
                      "gradonacelnik": "Ivan Radić",
                      "brojstanovnika": 97846,
                      "nadmorskavisina": 90,
                      "links": [
                        {
                          "href": "/api/grad/4",
                          "rel": "self",
                          "type": "GET"
                        },
                        {
                          "href": "/api/grad/ime/Osijek",
                          "rel": "self",
                          "type": "GET"
                        },
                        {
                          "href": "/api/gradovi/populacija/97845",
                          "rel": "related",
                          "type": "GET"
                        }
                      ]
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "status": "OK",
                  "message": "Unable to retrieve data from database",
                  "response": null
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Dodavanje novog grada",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "https://github.com/Luka147m/OR-labos/blob/main/backend/addRequestSchema.json",
                "properties": {
                  "imegrada": {
                    "title": "imegrada",
                    "description": "Sadrži ime grada",
                    "type": "string"
                  },
                  "zupanija": {
                    "title": "zupanija",
                    "description": "Županija u kojoj se nalazi grad",
                    "type": "string"
                  },
                  "gradonacelnik": {
                    "title": "gradonacelnik",
                    "description": "Trenutni gradonačelnik",
                    "type": "string"
                  },
                  "brojstanovnika": {
                    "title": "brojstanovnika",
                    "description": "Broj stanovnika pojedinog grada",
                    "type": "integer"
                  },
                  "povrsina": {
                    "title": "povrsina",
                    "description": "Površina pod vlasništvom grada (u kilometrima kvadratnim)",
                    "type": "number"
                  },
                  "godinaosnutka": {
                    "title": "godinaosnutka",
                    "description": "Godina kada je grad osnovan (negativna godina predstavlja godinu prije krista)",
                    "type": "integer"
                  },
                  "latitude": {
                    "title": "latitude",
                    "description": "Zemljopisna širina grada",
                    "type": "number"
                  },
                  "longitude": {
                    "title": "longitude",
                    "description": "Zemljopisna visina grada",
                    "type": "number"
                  },
                  "nadmorskavisina": {
                    "title": "nadmorskavisina",
                    "description": "Nadmorska visina na kojoj se nalazi grad",
                    "type": "integer"
                  },
                  "kvartovi": {
                    "title": "kvartovi",
                    "description": "Popis kvartova grada",
                    "type": "array",
                    "items": {
                      "title": "Kvart",
                      "description": "Objekt koji sadrži informacije o kvartu",
                      "type": "object",
                      "required": ["nazivkvarta", "brojkvartstan"],
                      "properties": {
                        "nazivkvarta": {
                          "title": "nazivkvarta",
                          "description": "Naziv kvarta",
                          "type": "string"
                        },
                        "brojkvartstan": {
                          "title": "brojkvartstan",
                          "description": "Broj stanovnika u pojedinom kvartu",
                          "type": "integer"
                        }
                      }
                    }
                  }
                },
                "required": [
                  "imegrada",
                  "zupanija",
                  "gradonacelnik",
                  "brojstanovnika",
                  "povrsina",
                  "godinaosnutka",
                  "latitude",
                  "longitude",
                  "nadmorskavisina",
                  "kvartovi"
                ]
              },
              "example": {
                "imegrada": "Zagreb",
                "zupanija": "Grad Zagreb",
                "gradonacelnik": "Tomislav Tomašević",
                "brojstanovnika": 777183,
                "povrsina": 641.2,
                "nadmorskavisina": 122,
                "godinaosnutka": 1094,
                "latitude": 45.813,
                "longitude": 15.9775,
                "kvartovi": [
                  {
                    "nazivkvarta": "Gornja Dubrava",
                    "brojkvartstan": 58908
                  },
                  {
                    "nazivkvarta": "Maksimir",
                    "brojkvartstan": 47989
                  },
                  {
                    "nazivkvarta": "Trešnjevka",
                    "brojkvartstan": 119475
                  }
                ]
              }
            }
          },
          "description": "Upit mora zadovoljavati <a href=\"https://github.com/Luka147m/OR-labos/blob/main/backend/addRequestSchema.json\">JSON schemu</a>"
        },
        "responses": {
          "200": {
            "description": "Uspješno dodan grad",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "status": "OK",
                  "message": "Entry added successfully with id: 1",
                  "response": [
                    {
                      "id": 1,
                      "imegrada": "Zagreb",
                      "kvartovi": [
                        {
                          "nazivkvarta": "Gornja Dubrava",
                          "brojkvartstan": 58908
                        },
                        {
                          "nazivkvarta": "Maksimir",
                          "brojkvartstan": 47989
                        },
                        {
                          "nazivkvarta": "Trešnjevka",
                          "brojkvartstan": 119475
                        }
                      ],
                      "latitude": 45.813,
                      "povrsina": 641.2,
                      "zupanija": "Grad Zagreb",
                      "longitude": 15.9775,
                      "godinaosnutka": 1094,
                      "gradonacelnik": "Tomislav Tomašević",
                      "brojstanovnika": 777183,
                      "nadmorskavisina": 122,
                      "links": [
                        {
                          "href": "/api/grad/1",
                          "rel": "self",
                          "type": "GET"
                        },
                        {
                          "href": "/api/grad/ime/Zagreb",
                          "rel": "self",
                          "type": "GET"
                        },
                        {
                          "href": "/api/gradovi/populacija/777182",
                          "rel": "related",
                          "type": "GET"
                        }
                      ]
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "status": "Internal Server Error",
                  "message": "Unable to retrieve data from database",
                  "response": null
                }
              }
            }
          }
        }
      }
    },
...
    "/api/grad/ime/{name}": {
      "get": {
        "summary": "Dohvaća podatke o gradu pomoću imena grada (ili dijela imena) u URL-u",
        "parameters": [
          {
            "name": "name",
            "description": "Ime traženog grada ili dio imena",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Uspješno dohvaćanje podataka o gradu",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "status": "OK",
                  "message": "Fetched all entries from database matching the name: Požega",
                  "response": [
                    {
                      "id": 6,
                      "imegrada": "Požega",
                      "kvartovi": [
                        {
                          "nazivkvarta": "Sajmište",
                          "brojkvartstan": 5432
                        }
                      ],
                      "latitude": 45.33,
                      "povrsina": 133.91,
                      "zupanija": "Požeško-slavonska",
                      "longitude": 17.67,
                      "godinaosnutka": 1100,
                      "gradonacelnik": "Željko Glavić",
                      "brojstanovnika": 22564,
                      "nadmorskavisina": 152,
                      "links": [
                        {
                          "href": "/api/grad/6",
                          "rel": "self",
                          "type": "GET"
                        },
                        {
                          "href": "/api/grad/ime/Požega",
                          "rel": "self",
                          "type": "GET"
                        },
                        {
                          "href": "/api/gradovi/populacija/22563",
                          "rel": "related",
                          "type": "GET"
                        }
                      ]
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Podatci o gradu s zadanim imenom nisu pronađeni",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "status": "Not Found",
                  "message": "Failed to find entry matching the name:",
                  "response": null
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "status": "Internal Server Error",
                  "message": "Unable to retrieve data from database",
                  "response": null
                }
              }
            }
          }
        }
      }
    },
    "/api/gradovi/populacija/{number}": {
      "get": {
        "summary": "Dohvaća gradove koji imaju populaciju veću od zadanog broja u URL-u",
        "parameters": [
          {
            "name": "number",
            "description": "Broj, traže se gradovi s većom populacijom od toga broja",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Uspješno dohvaćanje podataka o gradu",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "status": "OK",
                  "message": "Fetched all entries from database that have population bigger than: 500000",
                  "response": [
                    {
                      "id": 1,
                      "imegrada": "Zagreb",
                      "kvartovi": [
                        {
                          "nazivkvarta": "Gornja Dubrava",
                          "brojkvartstan": 58908
                        },
                        {
                          "nazivkvarta": "Maksimir",
                          "brojkvartstan": 47989
                        },
                        {
                          "nazivkvarta": "Trešnjevka",
                          "brojkvartstan": 119475
                        }
                      ],
                      "latitude": 45.813,
                      "povrsina": 641.2,
                      "zupanija": "Grad Zagreb",
                      "longitude": 15.9775,
                      "godinaosnutka": 1094,
                      "gradonacelnik": "Tomislav Tomašević",
                      "brojstanovnika": 777183,
                      "nadmorskavisina": 122,
                      "links": [
                        {
                          "href": "/api/grad/1",
                          "rel": "self",
                          "type": "GET"
                        },
                        {
                          "href": "/api/grad/ime/Zagreb",
                          "rel": "self",
                          "type": "GET"
                        }
                      ]
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Podatci o gradu s većom populacijom od zadane nisu pronađeni",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "status": "Not Found",
                  "message": "Failed to find entries that have population bigger than: 5000000",
                  "response": null
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "status": "Internal Server Error",
                  "message": "Unable to retrieve data from database",
                  "response": null
                }
              }
            }
          }
        }
      }
    },
...
  }
}